requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
databases
python-jose[cryptography]
passlib[bcrypt]
PyJWT
python-decouple
python-multipart
pymysql
passlib==1.7.4
bcrypt==3.2.2
python-dotenv
langchain-text-splitters
langchain-community
langgraph
langchain-openai
langchain-core
pydub
Pillow
pandas
pypdf
unstructured
networkx
openpyxl
docx2txt
oss2